requirements/requirements.txt
Django==1.8.3
django-celery==3.1.16
dj-static==0.0.6
psycopg2==2.6.1
djangorestframework==3.2.1
markdown==2.6.2
django-oauth-toolkit==0.9.0
dry-rest-permissions==0.1.0
django-cors-headers==1.1.0
django-debug-toolbar==1.3.2
django-storages==1.1.8
boto==2.38.0
Pillow==2.6.1
